Refaktorering av header
