+++ src/index.less
@@ -29,21 +29,16 @@ body.no-scroll {
     width: 100%;
 }
 
-.desktopmeny,
-.mobilmeny {
-    width: 100%;
-    background-color: white;
-}
-
-.desktopmeny {
-    height: 106px;
-    padding: 1rem 0;
-}
-
 .header-z-wrapper {
     z-index: 1000;
     position: relative;
     background-color: white;
+}
+
+.arbeidsflate {
+    @media (max-width: @mobilMaxWidth) {
+        display: none;
+    }
 }
 
 .site-skiplinks,
