+++ src/komponenter/header/header-regular/HeaderRegular.tsx
@@ -30,23 +30,4 @@ export const RegularHeader = () => {
             </div>
         </Fragment>
     );
-
-    // return (
-    //     <Fragment>
-    //         <Skiplinks />
-    //         <div className="media-sm-mobil">
-    //             <Sticky alwaysSticky={mobilMenyIsOpen}>
-    //                 <MobilMenylinje language={language} />
-    //             </Sticky>
-    //         </div>
-    //         <div className="media-tablet-desktop">
-    //             <div className="header-z-wrapper">
-    //                 {language === Language.NORSK && <Arbeidsflatemeny />}
-    //                 <Sticky>
-    //                     <DesktopMenylinje />
-    //                 </Sticky>
-    //             </div>
-    //         </div>
-    //     </Fragment>
-    // );
 };
