+++ src/komponenter/header/header-regular/Hovedmeny.tsx
@@ -1,0 +1,98 @@
+import React from 'react';
+import { useSelector } from 'react-redux';
+import { Status } from 'api/api';
+import { AppState } from 'store/reducers';
+import { getHovedmenyNode } from 'utils/meny-storage-utils';
+import EkspanderbarMeny from 'komponenter/header/header-regular/common/ekspanderbar-meny/EkspanderbarMeny';
+import Spinner from 'komponenter/header/header-regular/common/spinner/Spinner';
+import { KbNavMain } from 'utils/keyboard-navigation/useKbNavMain';
+import { HovedmenyKnapp } from 'komponenter/header/header-regular/common/knapper/hovedmeny-knapp/HovedmenyKnapp';
+import { gaEvent } from 'utils/google-analytics';
+import { GACategory } from 'utils/google-analytics';
+import { toggleHovedmeny } from 'store/reducers/dropdown-toggle-duck';
+import { useDispatch } from 'react-redux';
+import { HovedmenyMobil } from 'komponenter/header/header-regular/mobil/hovedmeny/HovedmenyMobil';
+import HovedmenyDesktop from 'komponenter/header/header-regular/desktop/hovedmeny/HovedmenyDesktop';
+
+const stateSelector = (state: AppState) => ({
+    arbeidsflate: state.arbeidsflate.status,
+    menyPunkter: state.menypunkt,
+    language: state.language.language,
+    isOpen: state.dropdownToggles.hovedmeny,
+    sokIsOpen: state.dropdownToggles.sok,
+});
+
+const classname = 'hovedmeny';
+export const hovedmenyKnappId = 'hovedmeny-knapp-id';
+
+type Props = {
+    kbNavMainState: KbNavMain;
+};
+
+export const Hovedmeny = ({ kbNavMainState }: Props) => {
+    const dispatch = useDispatch();
+    const { arbeidsflate, menyPunkter, language, isOpen } = useSelector(
+        stateSelector
+    );
+
+    const hovedmenyPunkter = getHovedmenyNode(
+        menyPunkter.data,
+        language,
+        arbeidsflate
+    );
+
+    // Hide empty menues
+    if (menyPunkter.status === Status.OK && !hovedmenyPunkter?.hasChildren) {
+        return null;
+    }
+
+    const toggleMenu = () => {
+        gaEvent({
+            context: arbeidsflate,
+            category: GACategory.Header,
+            action: `meny-${isOpen ? 'close' : 'open'}`,
+        });
+        dispatch(toggleHovedmeny());
+    };
+
+    const knapp = (
+        <HovedmenyKnapp
+            isOpen={isOpen}
+            onClick={toggleMenu}
+            id={hovedmenyKnappId}
+        />
+    );
+
+    const mobilMeny = <HovedmenyMobil />;
+
+    const desktopMeny = (
+        <HovedmenyDesktop
+            arbeidsflate={arbeidsflate}
+            isOpen={isOpen}
+            language={language}
+            menyPunkter={hovedmenyPunkter}
+            kbNavMainState={kbNavMainState}
+        />
+    );
+
+    return (
+        <EkspanderbarMeny
+            isOpen={isOpen}
+            menyKnapp={knapp}
+            classname={classname}
+            id={classname}
+        >
+            {menyPunkter.status === Status.OK ? (
+                <>
+                    {mobilMeny}
+                    {desktopMeny}
+                </>
+            ) : (
+                <Spinner
+                    tekstId={'meny-loading'}
+                    className={isOpen ? 'spinner-container--active' : ''}
+                />
+            )}
+        </EkspanderbarMeny>
+    );
+};
