--- src/komponenter/header/header-regular/desktop/DesktopMenylinje.less
@@ -1,95 +1,0 @@
-@import (reference) '../../../../styling-variabler.less';
-
-.desktopmeny {
-    @margin-mellom: 2rem;
-    @margin-mellom-sm: 0.5rem;
-    @maxheight-menuitem: 48px;
-
-    border-bottom: 2px solid @header-separator-farge;
-
-    .margin-left-section() {
-        margin-left: @margin-mellom;
-        @media (max-width: @smallDesktopLimitWidth) {
-            margin-left: @margin-mellom-sm;
-        }
-    }
-
-    .margin-right-section() {
-        margin-right: @margin-mellom;
-        @media (max-width: @smallDesktopLimitWidth) {
-            margin-right: @margin-mellom-sm;
-        }
-    }
-
-    &__elementer {
-        display: flex;
-        flex-direction: row;
-        align-items: center;
-        flex-wrap: wrap;
-        justify-content: space-between;
-        margin: 0 auto;
-        max-width: @desktopMaxWidth;
-    }
-
-    .nav-logo-lenke {
-        order: 1;
-
-        @padding: 8px;
-        width: 100px;
-        padding: @padding;
-
-        margin-left: calc(@margins-sides-desktop - @padding);
-        margin-right: -@padding;
-
-        @media (max-width: @smallDesktopLimitWidth) {
-            margin-left: calc(@margins-sides-desktop-small - @padding);
-        }
-        .nav-logo-fill {
-            fill: @nav-logo-rod;
-        }
-    }
-
-    .desktop-hovedmeny__knapp {
-        order: 2;
-        max-height: @maxheight-menuitem;
-        .margin-left-section();
-    }
-
-    .desktop-sok-dropdown__knapp {
-        order: 3;
-        max-height: @maxheight-menuitem;
-        .margin-left-section();
-    }
-
-    &__spacer {
-        order: 4;
-        flex-grow: 1;
-        flex-shrink: 0;
-    }
-
-    .varselbjelle__knapp {
-        order: 5;
-        .margin-right-section();
-        max-height: @maxheight-menuitem;
-    }
-
-    .desktop-minside-meny__knapp {
-        order: 6;
-        .margin-right-section();
-        max-height: @maxheight-menuitem;
-    }
-
-    .login-knapp-container {
-        order: 7;
-        max-height: @maxheight-menuitem;
-        margin-right: @margins-sides-desktop;
-        @media (max-width: @smallDesktopLimitWidth) {
-            margin-right: @margins-sides-desktop-small;
-        }
-    }
-
-    .ekspanderbar__container {
-        order: 8;
-        width: 100%;
-    }
-}
