+++ src/komponenter/header/header-regular/desktop/DesktopMenylinje.tsx
@@ -1,9 +1,8 @@
 import React from 'react';
 import BEMHelper from 'utils/bem';
 import { SokDropdown } from './sok/SokDropdown';
-import { HovedmenyDesktop } from './hovedmeny/HovedmenyDesktop';
 import { MinsideMeny } from './minside-meny/MinsideMeny';
-import { VarslerDropdown } from './varsler/VarslerDropdown';
+import { VarslerDropdown } from '../varsler/VarslerDropdown';
 import { useKbNavMain } from 'utils/keyboard-navigation/useKbNavMain';
 import { MenuValue } from 'utils/meny-storage-utils';
 import { useSelector } from 'react-redux';
@@ -11,7 +10,6 @@ import { AppState } from 'store/reducers';
 import LoggInnKnapp from 'komponenter/header/header-regular/common/knapper/logg-inn-knapp/LoggInnKnapp';
 import NavLogoLenke from 'komponenter/common/nav-logo/NavLogoLenke';
 import { GACategory } from 'utils/google-analytics';
-import './DesktopMenylinje.less';
 
 export const desktopHeaderLogoId = 'desktop-header-logo-id';
 export const desktopLoginKnappId = 'desktop-login-knapp';
@@ -41,7 +39,6 @@ const DesktopMenylinje = () => {
                     }}
                     id={desktopHeaderLogoId}
                 />
-                <HovedmenyDesktop kbNavMainState={kbNavMainState} />
                 <SokDropdown kbNavMainState={kbNavMainState} />
                 <span className={cls.element('spacer')} />
                 {visVarslerDropdown && (
