+++ src/komponenter/header/header-regular/desktop/hovedmeny/HovedmenyDesktop.tsx
@@ -1,33 +1,19 @@
 import React from 'react';
 import { useState } from 'react';
 import { useEffect } from 'react';
-import { useSelector } from 'react-redux';
-import { Status } from 'api/api';
-import { AppState } from 'store/reducers';
-import { getHovedmenyNode } from 'utils/meny-storage-utils';
-import EkspanderbarMeny from 'komponenter/header/header-regular/common/ekspanderbar-meny/EkspanderbarMeny';
-import { HovedmenyVisning } from './hovedmeny-visning/HovedmenyVisning';
-import Spinner from 'komponenter/header/header-regular/common/spinner/Spinner';
 import { KbNavMain } from 'utils/keyboard-navigation/useKbNavMain';
 import { KbNavConfig } from 'utils/keyboard-navigation/kb-navigation-setup';
 import { configForNodeGroup } from 'utils/keyboard-navigation/kb-navigation-setup';
 import { useKbNavSub } from 'utils/keyboard-navigation/useKbNavSub';
 import { KbNavGroup } from 'utils/keyboard-navigation/kb-navigation';
 import { matchMedia } from 'utils/match-media-polyfill';
-import { HovedmenyKnapp } from 'komponenter/header/header-regular/common/knapper/hovedmeny-knapp/HovedmenyKnapp';
 import { getHovedmenyMaxColsPerRow } from 'utils/keyboard-navigation/kb-navigation-setup';
-import { gaEvent } from 'utils/google-analytics';
-import { GACategory } from 'utils/google-analytics';
-import { toggleHovedmeny } from 'store/reducers/dropdown-toggle-duck';
-import { useDispatch } from 'react-redux';
-
-const stateSelector = (state: AppState) => ({
-    arbeidsflate: state.arbeidsflate.status,
-    menyPunkter: state.menypunkt,
-    language: state.language.language,
-    isOpen: state.dropdownToggles.hovedmeny,
-    sokIsOpen: state.dropdownToggles.sok,
-});
+import { MenuValue } from 'utils/meny-storage-utils';
+import { Language } from 'store/reducers/language-duck';
+import { MenyNode } from 'store/reducers/menu-duck';
+import { Toppseksjon } from 'komponenter/header/header-regular/desktop/hovedmeny/topp-seksjon/Toppseksjon';
+import { Hovedseksjon } from 'komponenter/header/header-regular/desktop/hovedmeny/hoved-seksjon/Hovedseksjon';
+import { Bunnseksjon } from 'komponenter/header/header-regular/desktop/hovedmeny/bunn-seksjon/Bunnseksjon';
 
 const classname = 'desktop-hovedmeny';
 export const desktopHovedmenyKnappId = 'desktop-hovedmeny-knapp-id';
@@ -36,21 +22,20 @@ const nodeGroup = KbNavGroup.Hovedmeny;
 const mqlScreenWidth = matchMedia('(min-width: 1024px)');
 
 type Props = {
+    arbeidsflate: MenuValue;
+    menyPunkter?: MenyNode;
+    language: Language;
+    isOpen: boolean;
     kbNavMainState: KbNavMain;
 };
 
-export const HovedmenyDesktop = ({ kbNavMainState }: Props) => {
-    const dispatch = useDispatch();
-    const { arbeidsflate, menyPunkter, language, isOpen } = useSelector(
-        stateSelector
-    );
-
-    const hovedmenyPunkter = getHovedmenyNode(
-        menyPunkter.data,
-        language,
-        arbeidsflate
-    );
-
+export const HovedmenyDesktop = ({
+    kbNavMainState,
+    arbeidsflate,
+    menyPunkter,
+    language,
+    isOpen,
+}: Props) => {
     const kbConfig = configForNodeGroup[nodeGroup];
     const [kbNavConfig, setKbNavConfig] = useState<KbNavConfig>(kbConfig);
     useKbNavSub(kbNavConfig, kbNavMainState, isOpen);
@@ -70,50 +55,26 @@ export const HovedmenyDesktop = ({ kbNavMainState }: Props) => {
 
     useEffect(() => {
         updateMaxCols();
-    }, [hovedmenyPunkter, arbeidsflate]);
+    }, [menyPunkter, arbeidsflate]);
 
-    // Hide empty menues
-    if (menyPunkter.status === Status.OK && !hovedmenyPunkter?.hasChildren) {
+    if (!menyPunkter) {
         return null;
     }
 
-    const toggleMenu = () => {
-        gaEvent({
-            context: arbeidsflate,
-            category: GACategory.Header,
-            action: `meny-${isOpen ? 'close' : 'open'}`,
-        });
-        dispatch(toggleHovedmeny());
-    };
-
-    const knapp = (
-        <HovedmenyKnapp
-            isOpen={isOpen}
-            onClick={toggleMenu}
-            hovedmenyClassname={classname}
-            id={desktopHovedmenyKnappId}
-        />
-    );
-
     return (
-        <EkspanderbarMeny
-            isOpen={isOpen}
-            menyKnapp={knapp}
-            classname={classname}
-            id={classname}
-        >
-            {menyPunkter.status === Status.OK ? (
-                <HovedmenyVisning
-                    classname={classname}
-                    arbeidsflate={arbeidsflate}
-                    language={language}
-                    menyLenker={hovedmenyPunkter}
-                    isOpen={isOpen}
-                />
-            ) : (
-                <Spinner tekstId={'meny-loading'} />
-            )}
-        </EkspanderbarMeny>
+        <div className={classname}>
+            <Toppseksjon classname={classname} />
+            <Hovedseksjon
+                menyLenker={menyPunkter}
+                classname={classname}
+                isOpen={isOpen}
+            />
+            <Bunnseksjon
+                classname={classname}
+                language={language}
+                arbeidsflate={arbeidsflate}
+            />
+        </div>
     );
 };
