+++ src/komponenter/header/header-regular/desktop/hovedmeny/bunn-seksjon/Bunnseksjon.less
@@ -1,4 +1,4 @@
-@import (reference) '../../../../../../../styling-variabler.less';
+@import (reference) '../../../../../../styling-variabler.less';
 
 .desktop-hovedmeny {
     @bunnseksjon-cols: 3;
