+++ src/komponenter/header/header-regular/desktop/hovedmeny/hoved-seksjon/Hovedseksjon.less
@@ -1,4 +1,4 @@
-@import (reference) '../../../../../../../styling-variabler.less';
+@import (reference) '../../../../../../styling-variabler.less';
 
 .desktop-hovedmeny {
     @hovedseksjon-margin: 1rem;
