+++ src/komponenter/header/header-regular/desktop/hovedmeny/topp-seksjon/Toppseksjon.less
@@ -1,4 +1,4 @@
-@import (reference) '../../../../../../../styling-variabler.less';
+@import (reference) '../../../../../../styling-variabler.less';
 
 .desktop-hovedmeny {
     &__topp-seksjon {
