+++ src/komponenter/header/header-regular/mobil/MobilMenylinje.less
@@ -5,7 +5,7 @@
         display: flex;
         flex-direction: row;
         justify-content: space-between;
-        min-height: 88px;
+        height: 88px;
         padding: 0.5rem 0.5rem 0.5rem 1rem;
         border-bottom: 2px solid @header-separator-farge;
     }
