+++ src/komponenter/header/header-regular/mobil/hovedmeny/HovedmenyMobil.tsx
@@ -1,16 +1,12 @@
 import React from 'react';
 import { AppState } from 'store/reducers';
 import { useDispatch, useSelector } from 'react-redux';
-import { Status } from 'api/api';
 import { getHovedmenyNode, getMinsideMenyNode } from 'utils/meny-storage-utils';
 import MobilVisningsmeny from './meny-dropdown/MobilVisningsmeny';
 import { GACategory, gaEvent } from 'utils/google-analytics';
 import { toggleUndermenyVisning } from 'store/reducers/dropdown-toggle-duck';
 import { toggleHovedmeny } from 'store/reducers/dropdown-toggle-duck';
 import { dataInitState } from 'store/reducers/menu-duck';
-import EkspanderbarMeny from 'komponenter/header/header-regular/common/ekspanderbar-meny/EkspanderbarMeny';
-import Spinner from 'komponenter/header/header-regular/common/spinner/Spinner';
-import { HovedmenyKnapp } from 'komponenter/header/header-regular/common/knapper/hovedmeny-knapp/HovedmenyKnapp';
 
 const stateSelector = (state: AppState) => ({
     meny: state.menypunkt,
@@ -25,7 +21,7 @@ const classname = 'mobilmeny';
 export const mobilHovedmenyKnappId = `${classname}-knapp-id`;
 export const mobilSokInputId = `${classname}-sok-input`;
 
-const HovedmenyMobil = () => {
+export const HovedmenyMobil = () => {
     const dispatch = useDispatch();
     const {
         meny,
@@ -41,11 +37,6 @@ const HovedmenyMobil = () => {
         language,
         arbeidsflate
     );
-
-    // Hide empty menues
-    if (meny.status === Status.OK && !hovedmenyPunkter?.hasChildren) {
-        return null;
-    }
 
     const menutoggle = () => {
         gaEvent({
@@ -66,49 +57,19 @@ const HovedmenyMobil = () => {
 
     const isOpen = hovedIsOpen || underIsOpen || varselIsOpen;
 
-    const menyKnapp = (
-        <HovedmenyKnapp
-            isOpen={hovedIsOpen}
-            onClick={hovedmenutoggle}
-            hovedmenyClassname={classname}
-            id={mobilHovedmenyKnappId}
+    return (
+        <MobilVisningsmeny
+            classname={classname}
+            menyLenker={hovedmenyPunkter || dataInitState}
+            minsideLenker={
+                getMinsideMenyNode(meny.data, language) || dataInitState
+            }
+            menuIsOpen={hovedIsOpen}
+            underMenuIsOpen={underIsOpen}
+            varslerIsOpen={varselIsOpen}
+            togglemenu={menutoggle}
+            togglehovedmenu={hovedmenutoggle}
+            lang={language}
         />
     );
-
-    const dropdownInnhold =
-        meny.status === Status.OK ? (
-            <MobilVisningsmeny
-                classname={classname}
-                menyLenker={hovedmenyPunkter || dataInitState}
-                minsideLenker={
-                    getMinsideMenyNode(meny.data, language) || dataInitState
-                }
-                menuIsOpen={hovedIsOpen}
-                underMenuIsOpen={underIsOpen}
-                varslerIsOpen={varselIsOpen}
-                togglemenu={menutoggle}
-                togglehovedmenu={hovedmenutoggle}
-                lang={language}
-            />
-        ) : (
-            <Spinner
-                tekstId={'meny-loading'}
-                className={isOpen ? 'spinner-container--active' : ''}
-            />
-        );
-
-    return (
-        <>
-            <EkspanderbarMeny
-                classname={classname}
-                isOpen={isOpen}
-                menyKnapp={menyKnapp}
-                id={classname}
-            >
-                {dropdownInnhold}
-            </EkspanderbarMeny>
-        </>
-    );
 };
-
-export default HovedmenyMobil;
