+++ src/komponenter/header/header-regular/varsler/varselvisning/Varselvisning.less
@@ -1,4 +1,4 @@
-@import (reference) '../../../../../../styling-variabler.less';
+@import (reference) '../../../../../styling-variabler.less';
 
 .varsler-visning {
     ul {
